show comments counts and child comment count
