+++ src/app/(pages)/post/[postId]/page.tsx
@@ -3,15 +3,13 @@ import type { Metadata, ResolvingMetadata } from "next";
 import React from "react";
 import PageWraper from "@/components/Wrapers/PageWraper";
 import "./index.scss";
-import Comments from "./comments";
-import Post from "./post";
-import FeedPost from "@/components/feedPost/feedPost";
 import { getPostsByPostId, getPostsForMeta } from "@/services/api/api";
 
 //markdown render
 import markdownToTxt from "markdown-to-txt";
 
 import ReactMarkdown from "react-markdown";
+import PageContainer from "./pageContainer";
 interface Iprops {
   params: any;
 }
@@ -21,11 +19,7 @@ export default async function PostPage({ params }: Iprops) {
   const postData = await getPostsByPostId(postId);
   return (
     <PageWraper hideRightPanel>
-      <main className='post_page'>
-        {/* <Post post={postData} /> */}
-        <FeedPost post={postData} />
-        <Comments postId={postId} />
-      </main>
+      <PageContainer postData={postData} />
     </PageWraper>
   );
 }
