+++ src/components/feedPost/index.scss
@@ -1,6 +1,6 @@
 @import "/src/styles";
 .postcard_container {
-  width: 98%;
+  width: 100%;
   padding: 24px 20px;
   height: max-content;
   border-radius: $base-border-radius-md;
